requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
contextlib2==21.6.0
idna==3.3
requests==2.28.0
schema==0.7.5
urllib3==1.26.9
python-dotenv==1.1.0
